+++ Solverz/solvers/daesolver/__init__.py
@@ -1,4 +1,5 @@
 from Solverz.solvers.daesolver.rodas import Rodas
+from Solverz.solvers.daesolver.radau import Radau
 from Solverz.solvers.daesolver.ode15s import ode15s
 from Solverz.solvers.daesolver.beuler import backward_euler
 from Solverz.solvers.daesolver.trapezoidal import implicit_trapezoid